requirements.txt
fastapi
jinja2
psycopg2-binary
pytest
python-dotenv
requests